Define tapping term length & disable hold on other key press
Tapping term configures how long a key needs to be held before it is considered
a hold rather than a tap. This is a dynamic binding expected to change regularly
as I seek out the sweet spot for my typing style and familiarity with mod-taps.

Disabling HOLD_ON_OTHER_KEY_PRESS is expected to be a permanent change, which
allows typing with key rolls on mod-tap keys without accidentally calling the
hold behaviour. It is typically off by default, but for some reason is enabled
in the Keychron Q1V2 config file, so needs to be overridden in the keymap
specific config file.