Caps Lock key when tapped, lctl when held
This gives an improvement to my previous approach of having Caps Lock mapped to
Control by the OS. Firstly, it means that Caps Lock is available for use when
required. More importantly, it means that the caps lock key can be used for
Control when using machines/OSes which don't have caps lock mapped to control.